threat-classifier/requirements.txt
google-adk>=0.1.0
google-genai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
google-cloud-logging